pyproject.toml
[project]
name = "posttrain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.3",
]

[project.scripts]
posttrain = "posttrain.cli:main"

[project.optional-dependencies]
training = [
    "accelerate>=1.12.0",
    "bitsandbytes>=0.49.1",
    "datasets>=4.5.0",
    "peft>=0.18.1",
    "transformers>=5.1.0",
    "trl>=0.28.0",
]

[dependency-groups]
dev = [
    "pyrefly>=0.52.0",
    "pytest>=9.0.2",
    "ruff>=0.15.1",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/posttrain"]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.pyrefly]
ignore-missing-imports = ["torch", "torch.nn", "torch.nn.functional", "transformers", "peft"]
